+++ src/fir/fir.c
@@ -43,7 +43,7 @@ void fir_init(struct fir *fir)
  */
 void fir_process(struct fir *fir, const int16_t *coeffs,
 		 const int16_t *input, int16_t *output,
-		 size_t length, int filterLength, int channels)
+		 size_t length, int filterLength, uint8_t channels)
 {
 	const int16_t *inputx = input;
 	int32_t acc;
